// src/AutocompleteSelect/types/theme.ts
export type Theme = {
  borderRadius: string
  colors: {
    background: string
    input: {
      background: string
    }
    border: string
    borderBottomOpen: string
    borderRadius: string
    color: string
    focused: {
      background: string
      color: string
    }
    selected: {
      background: string
      color: string
    }
    disabled: {
      background: string
      color: string
    }
  }
}